Stop using Kazan.Client altogether.
Stops us getting our own deprecation warnings on compile.
